push edges towards derived layers
