+++ include/spark_dsg/dynamic_scene_graph_layer.h
@@ -36,6 +36,7 @@
 #include <Eigen/Core>
 
 #include "spark_dsg/base_layer.h"
+#include "spark_dsg/edge_container.h"
 
 namespace spark_dsg {
