+++ include/spark_dsg/spark_dsg_fwd.h
@@ -5,6 +5,8 @@ namespace spark_dsg {
 class Mesh;
 struct NodeAttributes;
 struct EdgeAttributes;
+struct SceneGraphEdge;
+struct EdgeContainer;
 
 class DynamicSceneGraph;
 using DynamicSceneGraphPtr = std::shared_ptr<DynamicSceneGraph>;
